+++ src/wai/common/logging/__init__.py
@@ -2,12 +2,13 @@
 Package for logging utilities.
 """
 from ._LoggingMixin import LoggingMixin
-from ._standard_logger import (
+from ._standard_application_root_logger import (
     DEBUG_HANDLER_NAME,
     INFO_HANDLER_NAME,
     WARNING_HANDLER_NAME,
     ERROR_HANDLER_NAME,
     exact_level_filter,
     create_standard_handler,
-    create_standard_logger
+    create_standard_application_root_logger
 )
+from ._standard_library_root_logger import create_standard_library_root_logger